+++ front/src/components/header/DarkMode.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 
 const DarkMode = () => {
-  let clickedClass = "clicked";
   const body = document.body;
   const lightTheme = "light";
   const darkTheme = "is_dark";
@@ -16,19 +15,6 @@ const DarkMode = () => {
     body.classList.add(darkTheme);
   }
 
-  const switchTheme = (e) => {
-    if (theme === darkTheme) {
-      body.classList.replace(darkTheme, lightTheme);
-      e.target.classList.remove(clickedClass);
-      localStorage.setItem("theme", "light");
-      theme = lightTheme;
-    } else {
-      body.classList.replace(lightTheme, darkTheme);
-      e.target.classList.add(clickedClass);
-      localStorage.setItem("theme", "is_dark");
-      theme = darkTheme;
-    }
-  };
   return <></>;
 };
